remove deprecation warning

+++ rdkit/Chem/Graphs.py
@@ -1,6 +1,5 @@
-# $Id$
 #
-# Copyright (C) 2001-2008 greg landrum and rational discovery llc
+# Copyright (C) 2001-2022 Greg Landrum and other RDKit contributors
 #
 #   @@ All Rights Reserved @@
 #  This file is part of the RDKit.
@@ -40,7 +39,7 @@ def CharacteristicPolynomial(mol, mat=None):
         A = mat
     I = 1. * numpy.identity(nAtoms)
     An = A
-    res = numpy.zeros(nAtoms + 1, numpy.float)
+    res = numpy.zeros(nAtoms + 1, float)
     res[0] = 1.0
     for n in range(1, nAtoms + 1):
         res[n] = 1. / n * numpy.trace(An)
